Fix crash "int is not a subtype of double?" su posizionamento widget
Le costanti di dimensione di default (immagine, lottie, divider, sized box,
constrained box, progress, page view, rating bar) erano dichiarate come
letterali int ma castate `as double?` nei fromJson: `140 as double?` crasha a
runtime. Rese tutte letterali double, così il cast diventa superfluo.

- Costanti DEFAULT_* portate a double in comman_property_view.dart
- Rimossi 50 cast `as double`/`as double?` ora ridondanti (8 widget class)
- Allineati i fallback dei property view a .toInt().toString()
- Rimosse le costanti morte DEFAULT_LIMITED_BOX_HEIGHT/WIDTH

flutter analyze lib/ -> 0 error, 0 unnecessary_cast; flutter test -> 7/7.

+++ lib/widgetsProperty/divider_property_view.dart
@@ -166,7 +166,7 @@ class DividerPropertyViewState extends State<DividerPropertyView> {
               width: widthPropertySize,
               child: getTextField(
                 controller: TextEditingController(
-                  text: dividerClass.height != null ? dividerClass.height.toString() : DEFAULT_DIVIDER_HEIGHT.toString(),
+                  text: dividerClass.height != null ? dividerClass.height.toString() : DEFAULT_DIVIDER_HEIGHT.toInt().toString(),
                 ),
                 textAlign: TextAlign.center,
                 inputFormatter: <TextInputFormatter>[
